Implementing All Sorts to GUI.py
Confirmed that Filter By with Sorts are working correctly
