+++ GUI.py
@@ -3,10 +3,12 @@ from tkinter import ttk
 from tkinter import messagebox
 from GameList import Game, games_data, GameGraph,get_games
 import random
+from Sort import merge_sort, quick_sort, selection_sort, bubble_sort, insertion_sort, heap_sort
 
 # Constant for color theme
 BLUE = "#1f6aa5"
 DARK = "gray14"
+
 class Node:
     def __init__(self, value):
         self.value = value
@@ -70,7 +72,8 @@ class GameShopGUI:
         self.game_graph = None
         self.queue = Queue()
         self.total_customers = 25
-        self.your_position = 10
+        self.your_position = 3
+        self.games_data = games_data
 
         # Create banner label spanning across the top
         self.banner_label = tk.Label(master, text="Welcome to GameShop", bg=BLUE, fg="white", font=("Helvetica", 24))
@@ -107,15 +110,86 @@ class GameShopGUI:
         # Start updating the queue label and countdown
         update_queue_label()
 
-    def sort_games(self, criterion="title", descending=False):
+    def add_to_cart(self, game_data):
+        """Add a new game to the cart"""
+        self.cart.push(game_data)
+        self.last_added_genre = game_data['genre'] 
+        self.update_cart_display()
+        self.update_recommended_games()
+
+    # Sorts Game based on Filter (Utilizes 6 Different Sorts)
+    def sort_games(self, sort_criteria, order="ascending"):
         """Sorts the list of games based on a given criterion."""
-        if criterion == "title":
-            sorted_games = sorted(games_data, key=lambda x: x[criterion], reverse=descending)
-        elif criterion in ["price", "review"]:
-            sorted_games = sorted(games_data, key=lambda x: float(x[criterion]), reverse=descending)
-        else:
-            sorted_games = games_data
+        games = self.games_data
+
+        if sort_criteria == 'title':
+            key_func = lambda game: game[sort_criteria]
+            if order == "ascending":
+                sorted_games = merge_sort(games, key=key_func)
+            else:
+                sorted_games = quick_sort(games, key=key_func)
+
+        elif sort_criteria == 'price':
+            key_func = lambda game: game[sort_criteria]
+            if order == "ascending":
+                sorted_games = insertion_sort(games, key=key_func)
+            else:
+                sorted_games = heap_sort(games, key=key_func)
+
+        elif sort_criteria == 'review':
+            key_func = lambda game: game[sort_criteria]
+            if order == "ascending":
+                sorted_games = bubble_sort(games, key=key_func)
+            else:
+                sorted_games = selection_sort(games, key=key_func)
+
         return sorted_games
+
+    def display_games(self, sorted_games_data):
+        # Clear existing game display from the inner_frame
+        for widget in self.inner_frame.winfo_children():
+            widget.destroy()
+
+        # Add column headers
+        headers = ["Title", "Price", "Review Score", "Genre", "ESRB Rating"]
+        for idx, header in enumerate(headers):
+            header_label = tk.Label(self.inner_frame, text=header, font=("Helvetica", 14))
+            header_label.grid(row=0, column=idx, padx=10)
+
+        # Repopulate the game display with sorted_games_data
+        for idx, game_data in enumerate(sorted_games_data, start=1):
+            game = Game(**game_data)  # Assuming each item in sorted_games_data is a dictionary
+
+            # Display Game Title, Price, Review Score, Genre, and ESRB Rating
+            tk.Label(self.inner_frame, text=game.title, font=("Helvetica", 14)).grid(row=idx, column=0, pady=(10, 0))
+            tk.Label(self.inner_frame, text=f"${game.price}", font=("Helvetica", 14)).grid(row=idx, column=1, pady=(10, 0))
+            tk.Label(self.inner_frame, text=f"{game.review} / 5", font=("Helvetica", 14)).grid(row=idx, column=2, pady=(10, 0))
+            tk.Label(self.inner_frame, text=game.genre, font=("Helvetica", 14)).grid(row=idx, column=3, pady=(10, 0))
+            tk.Label(self.inner_frame, text=game.esrb_rating, font=("Helvetica", 14)).grid(row=idx, column=4, pady=(10, 0))
+
+            # Add to Cart button
+            add_to_cart_button = tk.Button(self.inner_frame, text="Add to Cart", bg=BLUE, command=lambda gd=game_data: self.add_to_cart(gd))
+            add_to_cart_button.grid(row=idx, column=5, padx=10, pady=(10, 0))
+
+        self.canvas.configure(scrollregion=self.canvas.bbox("all"))
+
+    def on_filter_selection(self, event):
+        selected_option = event.widget.get()  # Get the selected filter option
+
+        # Mapping from the filter option to the sorting criteria and order
+        option_mapping = {
+            "Alphabetically: A-Z": ("title", "ascending"),
+            "Alphabetically: Z-A": ("title", "descending"),
+            "Price: High to Low": ("price", "descending"),
+            "Price: Low to High": ("price", "ascending"),
+            "Review: High to Low": ("review", "descending"),
+            "Review: Low to High": ("review", "ascending")
+        }
+
+        if selected_option in option_mapping:
+            sort_criteria, order = option_mapping[selected_option]
+            sorted_games_data = self.sort_games(sort_criteria, order)
+            self.display_games(sorted_games_data)
 
     def show_game_info_page(self, queue_window):
         """Show the game information"""
@@ -127,12 +201,12 @@ class GameShopGUI:
         game_info_window.geometry(f"{width}x{height}")
 
         # Create a frame to contain the game information and filter selection box
-        frame = tk.Frame(game_info_window)
-        frame.pack(fill=tk.BOTH, expand=True, padx=20, pady=20)
-        frame.grid_rowconfigure(0, weight = 1)
-        frame.grid_columnconfigure(2, weight = 1)
+        self.frame = tk.Frame(game_info_window)
+        self.frame.pack(fill=tk.BOTH, expand=True, padx=20, pady=20)
+        self.frame.grid_rowconfigure(0, weight = 1)
+        self.frame.grid_columnconfigure(2, weight = 1)
         # Add filter selection box on the left
-        filter_label = tk.Label(frame, text="Filter by:", font=("Helvetica", 12))
+        filter_label = tk.Label(self.frame, text="Filter by:", font=("Helvetica", 12))
         filter_label.grid(row=0, column=0, padx=(0, 5), sticky="e")
         filter_options = [
             "Alphabetically: A-Z",
@@ -142,73 +216,65 @@ class GameShopGUI:
             "Review: High to Low",
             "Review: Low to High"
         ]
-        filter_combobox = ttk.Combobox(frame, values=filter_options, state="readonly", width=20)
+        filter_combobox = ttk.Combobox(self.frame, values=filter_options, state="readonly", width=20)
         filter_combobox.current(0)  # Select the first option by default
         filter_combobox.grid(row=0, column=1, padx=(0, 10), sticky="w")
+        filter_combobox.bind("<<ComboboxSelected>>", self.on_filter_selection)
 
         # Create a Canvas widget for scrolling
-        canvas = tk.Canvas(frame)
-        canvas.grid(row=0, column=2, rowspan=2, sticky="nsew")
+        self.canvas = tk.Canvas(self.frame)
+        self.canvas.grid(row=0, column=2, rowspan=2, sticky="nsew")
 
         # Create a scrollbar for the canvas
-        v_scroll = tk.Scrollbar(frame, orient="vertical", command=canvas.yview)
+        v_scroll = tk.Scrollbar(self.frame, orient="vertical", command=self.canvas.yview)
         v_scroll.grid(row=0, column=3, rowspan=2, sticky="ns")
 
         # Configure canvas to use scrollbar
-        canvas.configure(yscrollcommand=v_scroll.set)
-        canvas.bind("<Configure>", lambda e: canvas.configure(scrollregion=canvas.bbox("all")))
+        self.canvas.configure(yscrollcommand=v_scroll.set)
+        self.canvas.bind("<Configure>", lambda e: self.canvas.configure(scrollregion=self.canvas.bbox("all")))
 
         # Create a frame inside the canvas
-        inner_frame = tk.Frame(canvas)
-        canvas.create_window((0, 0), window=inner_frame, anchor="nw")
+        self.inner_frame = tk.Frame(self.canvas)
+        self.canvas.create_window((0, 0), window=self.inner_frame, anchor="nw")
 
         def UpdateScroll(event):
             """Update the scroll bar after resizing """
-            canvas.configure(scrollregion=canvas.bbox("all"))
+            self.canvas.configure(scrollregion=self.canvas.bbox("all"))
         
-        inner_frame.bind("<Configure>",UpdateScroll)
+        self.inner_frame.bind("<Configure>",UpdateScroll)
 
         # Add column headers
         headers = ["Title", "Price", "Review Score", "Genre", "ESRB Rating"]
         for idx, header in enumerate(headers):
-            header_label = tk.Label(inner_frame, text=header, font=("Helvetica", 14))
+            header_label = tk.Label(self.inner_frame, text=header, font=("Helvetica", 14))
             header_label.grid(row=0, column=idx, padx=10)
 
-        def add_to_cart(game_data):
-            """Add a new game to the cart"""
-            self.cart.push(game_data)
-            self.last_added_genre = game_data['genre'] 
-            self.update_cart_display()
-            self.update_recommended_games()
-
         # Display sorted game data
-        sorted_games_data = self.sort_games("title", False)
+        sorted_games_data = self.sort_games("title", "ascending")
         for idx, game_data in enumerate(sorted_games_data, start=1):
             game = Game(**game_data)
 
             # Display Game Title, Price, Review Score, Genre, and ESRB Rating
-            tk.Label(inner_frame, text=game.title, font=("Helvetica", 14), anchor="center").grid(row=idx, column=0, pady=(10, 0), sticky= "nsew")
-            tk.Label(inner_frame, text=f"${game.price}", font=("Helvetica", 14), anchor="center").grid(row=idx, column=1, pady=(10, 0), sticky= "nsew")
-            tk.Label(inner_frame, text=f"{game.review} / 5", font=("Helvetica", 14), anchor="center").grid(row=idx, column=2, pady=(10, 0), sticky= "nsew")
-            tk.Label(inner_frame, text=game.genre, font=("Helvetica", 14), anchor="center").grid(row=idx, column=3, pady=(10, 0), sticky= "nsew")
-            tk.Label(inner_frame, text=game.esrb_rating, font=("Helvetica", 14), anchor="center").grid(row=idx, column=4, pady=(10, 0), sticky= "nsew")
+            tk.Label(self.inner_frame, text=game.title, font=("Helvetica", 14), anchor="center").grid(row=idx, column=0, pady=(10, 0), sticky= "nsew")
+            tk.Label(self.inner_frame, text=f"${game.price}", font=("Helvetica", 14), anchor="center").grid(row=idx, column=1, pady=(10, 0), sticky= "nsew")
+            tk.Label(self.inner_frame, text=f"{game.review} / 5", font=("Helvetica", 14), anchor="center").grid(row=idx, column=2, pady=(10, 0), sticky= "nsew")
+            tk.Label(self.inner_frame, text=game.genre, font=("Helvetica", 14), anchor="center").grid(row=idx, column=3, pady=(10, 0), sticky= "nsew")
+            tk.Label(self.inner_frame, text=game.esrb_rating, font=("Helvetica", 14), anchor="center").grid(row=idx, column=4, pady=(10, 0), sticky= "nsew")
         
             # Add to Cart button
-
-            add_to_cart_button = tk.Button(inner_frame, text="Add to Cart", bg=BLUE, command=lambda gd=game_data: add_to_cart(gd))
+            add_to_cart_button = tk.Button(self.inner_frame, text="Add to Cart", bg=BLUE, command=lambda gd=game_data: self.add_to_cart(gd))
             add_to_cart_button.grid(row=idx, column=5, padx=10, pady=(10, 0))
 
-
-        canvas.configure(scrollregion=canvas.bbox("all"))
+        self.canvas.configure(scrollregion=self.canvas.bbox("all"))
 
         # Add the canvas to the window using grid 
-        canvas.grid(row=0, column=2, rowspan=2, sticky="nsew")
+        self.canvas.grid(row=0, column=2, rowspan=2, sticky="nsew")
 
         # Add padding to the bottom for better spacing
-        tk.Label(inner_frame, text="", font=("Helvetica", 14)).grid(row=len(sorted_games_data) + 1, column=0, pady=(10, 20))
+        tk.Label(self.inner_frame, text="", font=("Helvetica", 14)).grid(row=len(sorted_games_data) + 1, column=0, pady=(10, 20))
 
         # Bind the inner frame to the canvas
-        inner_frame.bind("<Configure>", lambda e: canvas.configure(scrollregion=canvas.bbox("all")))
+        self.inner_frame.bind("<Configure>", lambda e: self.canvas.configure(scrollregion=self.canvas.bbox("all")))
 
         Bottom_frame = tk.Frame(game_info_window)
         Bottom_frame.pack(side = tk.BOTTOM, anchor= 'w', padx=10, pady=10)
@@ -222,7 +288,6 @@ class GameShopGUI:
         self.recommended_listbox.pack(side=tk.LEFT, padx=5, pady =15)  
         self.update_recommended_games()
 
-     
         # Initialize the cart Listbox for displaying and selecting cart items
         self.cart_listbox = tk.Listbox(Bottom_frame, height=10, width=50)
         self.cart_listbox.pack(side=tk.LEFT, padx=25)
